To fix a bug with weights
If weights=1/n for instance, the results are different (beta, b0)
compared with weights=1. After adding the following line
weights <- weights/sum(weights)
the beta, b0 are the same when weights=1/n or 1. The true issue seems to
related to how weights and response variables are treated in glmreg_fit
function when standardize=FALSE. That needs to be investiaged further in
the future.
 Changes to be committed:
	modified:   R/zipath.R